CVE_scanner/requirements.txt
requests
python-nmap
tabulate
